test(render): move worker tests beside implementation
