--- src/app/tests/mod.rs
@@ -1,1 +1,0 @@
-mod runtime_worker;
